Fix outOfResourcesNDRange runfail on x86 win debug
sprintf_s: Use %llu instead of %ld to format cl_ulong.

Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/192596
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>

+++ opencl-intel/tests/framework_test_type/outOfResourcesNDRange.cpp
@@ -46,7 +46,8 @@ cl_program buildProgram(cl_context clContext, cl_device_id clDevice, const char 
         return NULL;
 
     char buildOptions[1024];
-    SPRINTF_S(buildOptions, 1024, "-cl-opt-disable -DSIZE_OF_LOCAL=%ld -DSIZE_OF_PRIVATE=%ld", localSize, privateSize);
+    SPRINTF_S(buildOptions, 1024, "-cl-opt-disable -DSIZE_OF_LOCAL=%llu -DSIZE_OF_PRIVATE=%llu",
+        (unsigned long long)localSize, (unsigned long long)privateSize);
 
     printf("Build options <%s>, source:%s\n", buildOptions, programSrc);
     // Build program executable from source. Prevent any optimizations (we want big memory etc.)
